Update ES256K->ES256 signatures
